refactor: objects.py: moved hash_object from hash_object.py
