+++ util/__init__.py
@@ -1,7 +1,6 @@
 from util.cat_file import cat_file
 from util.detect_dot_git import detect_dot_git
 from util.get_path_in_repo import get_path_in_repo
-from util.hash_object import hash_object
 from util.heads import (
     get_current_branch,
     read_branch,
@@ -9,7 +8,7 @@ from util.heads import (
     write_branch,
     write_head_hash,
 )
-from util.objects import read_object, write_object
+from util.objects import hash_object, read_object, write_object
 from util.read_index import read_index
 from util.tree import parse_tree, read_tree, write_tree
 from util.write_index import write_index
